Refactors files
